Large Infra: Jest Setup with base React components.
This is a bit of a large change.

I first wanted to get jest up and running but also have a setup for this project that wouldnt be a pain to work with as I develop foundation during the evenings and weekends.

We pretty much have the basic tools necessary to begin adding static, unit and integration tests in place.  Not only that, the application comes packed with helpful scripts for cleaning up ones codebase on the go, as well as testing it.  This way we are not relying on our CI to tell us that a build is failing or a change request should I bring someone on board in the future for this.