Infra: added typeahead support to jest.
I wanted to improve the test watch experience a little, without having to rely on having tools specifically on my system.  jest-watch-typeahead is nice because it lets you know which of your tests will be ran based on your filter.

Not only that you can toggle down and select a specific test if found.